+++ CHANGELOG.md
@@ -22,6 +22,7 @@ It is based on Tidepool Blip 1.27.
 - YLP-1082 Improve daily view rendering
 - YLP-1114 Review routes and data structure for message-api
 - YLP-1172 Cannot debug unit test of blip (front)
+- YLP-1174 Removing add patient functionality for a caregiver (not used)
 
 ## 2.0.5 - 2021-12-06
 ### Improvements